separate Database into interface and implementation
