+++ ktorm-core/src/main/kotlin/org/ktorm/database/Database.kt
@@ -1,19 +1,3 @@
-/*
- * Copyright 2018-2021 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.ktorm.database
 
 import org.ktorm.dsl.Query
@@ -25,11 +9,14 @@ import org.ktorm.logging.detectLoggerImplementation
 import org.springframework.dao.DataAccessException
 import org.springframework.jdbc.support.SQLErrorCodeSQLExceptionTranslator
 import org.springframework.transaction.annotation.Transactional
-import java.sql.*
+import java.sql.Connection
+import java.sql.DriverManager
+import java.sql.SQLException
 import java.util.*
 import javax.sql.DataSource
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
+import kotlin.contracts.InvocationKind.EXACTLY_ONCE
 import kotlin.contracts.contract
 
 /**
@@ -99,34 +86,33 @@ import kotlin.contracts.contract
  * ```
  * More details about SQL DSL, see [Query], about sequence APIs, see [EntitySequence].
  */
-public class Database(
-
+public interface Database {
     /**
      * The transaction manager used to manage connections and transactions.
      */
-    public val transactionManager: TransactionManager,
+    public val transactionManager: TransactionManager
 
     /**
      * The dialect, auto detects an implementation by default using JDK [ServiceLoader] facility.
      */
-    public val dialect: SqlDialect = detectDialectImplementation(),
+    public val dialect: SqlDialect
 
     /**
      * The logger used to output logs, auto detects an implementation by default.
      */
-    public val logger: Logger = detectLoggerImplementation(),
+    public val logger: Logger
 
     /**
      * Function used to translate SQL exceptions so as to rethrow them to users.
      */
-    public val exceptionTranslator: ((SQLException) -> Throwable)? = null,
+    public val exceptionTranslator: ((SQLException) -> Throwable)?
 
     /**
      * Whether we need to always quote SQL identifiers in the generated SQLs.
      *
      * @since 3.1.0
      */
-    public val alwaysQuoteIdentifiers: Boolean = false,
+    public val alwaysQuoteIdentifiers: Boolean
 
     /**
      * Whether we need to output the generated SQLs in upper case.
@@ -135,8 +121,8 @@ public class Database(
      *
      * @since 3.1.0
      */
-    public val generateSqlInUpperCase: Boolean? = null
-) {
+    public val generateSqlInUpperCase: Boolean?
+
     /**
      * The URL of the connected database.
      */
@@ -244,113 +230,6 @@ public class Database(
      */
     public val maxColumnNameLength: Int
 
-    init {
-        fun Result<String?>.orEmpty() = getOrNull().orEmpty()
-        fun Result<Boolean>.orFalse() = getOrDefault(false)
-
-        useConnection { conn ->
-            val metadata = conn.metaData
-            url = metadata.runCatching { url }.orEmpty()
-            name = url.substringAfterLast('/').substringBefore('?')
-            productName = metadata.runCatching { databaseProductName }.orEmpty()
-            productVersion = metadata.runCatching { databaseProductVersion }.orEmpty()
-            keywords = ANSI_SQL_2003_KEYWORDS + metadata.runCatching { sqlKeywords }.orEmpty().toUpperCase().split(',')
-            identifierQuoteString = metadata.runCatching { identifierQuoteString }.orEmpty().trim()
-            extraNameCharacters = metadata.runCatching { extraNameCharacters }.orEmpty()
-            supportsMixedCaseIdentifiers = metadata.runCatching { supportsMixedCaseIdentifiers() }.orFalse()
-            storesMixedCaseIdentifiers = metadata.runCatching { storesMixedCaseIdentifiers() }.orFalse()
-            storesUpperCaseIdentifiers = metadata.runCatching { storesUpperCaseIdentifiers() }.orFalse()
-            storesLowerCaseIdentifiers = metadata.runCatching { storesLowerCaseIdentifiers() }.orFalse()
-            supportsMixedCaseQuotedIdentifiers = metadata.runCatching { supportsMixedCaseQuotedIdentifiers() }.orFalse()
-            storesMixedCaseQuotedIdentifiers = metadata.runCatching { storesMixedCaseQuotedIdentifiers() }.orFalse()
-            storesUpperCaseQuotedIdentifiers = metadata.runCatching { storesUpperCaseQuotedIdentifiers() }.orFalse()
-            storesLowerCaseQuotedIdentifiers = metadata.runCatching { storesLowerCaseQuotedIdentifiers() }.orFalse()
-            maxColumnNameLength = metadata.runCatching { maxColumnNameLength }.getOrDefault(0)
-        }
-
-        if (logger.isInfoEnabled()) {
-            val msg = "Connected to %s, productName: %s, productVersion: %s, logger: %s, dialect: %s"
-            logger.info(msg.format(url, productName, productVersion, logger, dialect))
-        }
-    }
-
-    /**
-     * Obtain a connection and invoke the callback function with it.
-     *
-     * If the current thread has opened a transaction, then this transaction's connection will be used.
-     * Otherwise, Ktorm will pass a new-created connection to the function and auto close it after it's
-     * not useful anymore.
-     *
-     * @param func the executed callback function.
-     * @return the result of the callback function.
-     */
-    @OptIn(ExperimentalContracts::class)
-    public inline fun <T> useConnection(func: (Connection) -> T): T {
-        contract {
-            callsInPlace(func, InvocationKind.EXACTLY_ONCE)
-        }
-
-        try {
-            val transaction = transactionManager.currentTransaction
-            val connection = transaction?.connection ?: transactionManager.newConnection()
-
-            try {
-                return func(connection)
-            } finally {
-                if (transaction == null) connection.close()
-            }
-        } catch (e: SQLException) {
-            throw exceptionTranslator?.invoke(e) ?: e
-        }
-    }
-
-    /**
-     * Execute the specific callback function in a transaction and returns its result if the execution succeeds,
-     * otherwise, if the execution fails, the transaction will be rollback.
-     *
-     * Note:
-     *
-     * - Any exceptions thrown in the callback function can trigger a rollback.
-     * - This function is reentrant, so it can be called nested. However, the inner calls don’t open new transactions
-     * but share the same ones with outers.
-     * - Since version 3.3.0, the default isolation has changed to null (stands for the default isolation level of the
-     * underlying datastore), not [TransactionIsolation.REPEATABLE_READ] anymore.
-     *
-     * @param isolation transaction isolation, null for the default isolation level of the underlying datastore.
-     * @param func the executed callback function.
-     * @return the result of the callback function.
-     */
-    @OptIn(ExperimentalContracts::class)
-    public inline fun <T> useTransaction(isolation: TransactionIsolation? = null, func: (Transaction) -> T): T {
-        contract {
-            callsInPlace(func, InvocationKind.EXACTLY_ONCE)
-        }
-
-        val current = transactionManager.currentTransaction
-        val isOuter = current == null
-        val transaction = current ?: transactionManager.newTransaction(isolation)
-        var throwable: Throwable? = null
-
-        try {
-            return func(transaction)
-        } catch (e: SQLException) {
-            throwable = exceptionTranslator?.invoke(e) ?: e
-            throw throwable
-        } catch (e: Throwable) {
-            throwable = e
-            throw throwable
-        } finally {
-            if (isOuter) {
-                @Suppress("ConvertTryFinallyToUseCall")
-                try {
-                    if (throwable == null) transaction.commit() else transaction.rollback()
-                } finally {
-                    transaction.close()
-                }
-            }
-        }
-    }
-
     /**
      * Format the specific [SqlExpression] to an executable SQL string with execution arguments.
      *
@@ -363,43 +242,7 @@ public class Database(
         expression: SqlExpression,
         beautifySql: Boolean = false,
         indentSize: Int = 2
-    ): Pair<String, List<ArgumentExpression<*>>> {
-        val formatter = dialect.createSqlFormatter(this, beautifySql, indentSize)
-        formatter.visit(expression)
-        return Pair(formatter.sql, formatter.parameters)
-    }
-
-    /**
-     * Format the given [expression] to a SQL string with its execution arguments, then create
-     * a [PreparedStatement] from the this database using the SQL string and execute the specific
-     * callback function with it. After the callback function completes, the statement will be
-     * closed automatically.
-     *
-     * @since 2.7
-     * @param expression the SQL expression to be executed.
-     * @param func the callback function.
-     * @return the result of the callback function.
-     */
-    @OptIn(ExperimentalContracts::class)
-    public inline fun <T> executeExpression(expression: SqlExpression, func: (PreparedStatement) -> T): T {
-        contract {
-            callsInPlace(func, InvocationKind.EXACTLY_ONCE)
-        }
-
-        val (sql, args) = formatExpression(expression)
-
-        if (logger.isDebugEnabled()) {
-            logger.debug("SQL: $sql")
-            logger.debug("Parameters: " + args.map { "${it.value}(${it.sqlType.typeName})" })
-        }
-
-        useConnection { conn ->
-            conn.prepareStatement(sql).use { statement ->
-                statement.setArguments(args)
-                return func(statement)
-            }
-        }
-    }
+    ): Pair<String, List<ArgumentExpression<*>>>
 
     /**
      * Format the given [expression] to a SQL string with its execution arguments, then execute it via
@@ -409,19 +252,7 @@ public class Database(
      * @param expression the SQL expression to be executed.
      * @return the result [CachedRowSet].
      */
-    public fun executeQuery(expression: SqlExpression): CachedRowSet {
-        executeExpression(expression) { statement ->
-            statement.executeQuery().use { rs ->
-                val rowSet = CachedRowSet(rs)
-
-                if (logger.isDebugEnabled()) {
-                    logger.debug("Results: ${rowSet.size()}")
-                }
-
-                return rowSet
-            }
-        }
-    }
+    public fun executeQuery(expression: SqlExpression): CachedRowSet
 
     /**
      * Format the given [expression] to a SQL string with its execution arguments, then execute it via
@@ -431,17 +262,7 @@ public class Database(
      * @param expression the SQL expression to be executed.
      * @return the effected row count.
      */
-    public fun executeUpdate(expression: SqlExpression): Int {
-        executeExpression(expression) { statement ->
-            val effects = statement.executeUpdate()
-
-            if (logger.isDebugEnabled()) {
-                logger.debug("Effects: $effects")
-            }
-
-            return effects
-        }
-    }
+    public fun executeUpdate(expression: SqlExpression): Int
 
     /**
      * Format the given [expression] to a SQL string with its execution arguments, execute it via
@@ -451,22 +272,7 @@ public class Database(
      * @param expression the SQL expression to be executed.
      * @return a [Pair] combines the effected row count and the generated keys.
      */
-    public fun executeUpdateAndRetrieveKeys(expression: SqlExpression): Pair<Int, CachedRowSet> {
-        val (sql, args) = formatExpression(expression)
-
-        if (logger.isDebugEnabled()) {
-            logger.debug("SQL: $sql")
-            logger.debug("Parameters: " + args.map { "${it.value}(${it.sqlType.typeName})" })
-        }
-
-        val (effects, rowSet) = dialect.executeUpdateAndRetrieveKeys(this, sql, args)
-
-        if (logger.isDebugEnabled()) {
-            logger.debug("Effects: $effects")
-        }
-
-        return Pair(effects, rowSet)
-    }
+    public fun executeUpdateAndRetrieveKeys(expression: SqlExpression): Pair<Int, CachedRowSet>
 
     /**
      * Batch execute the given SQL expressions and return the effected row counts for each expression.
@@ -478,41 +284,7 @@ public class Database(
      * @param expressions the SQL expressions to be executed.
      * @return the effected row counts for each sub-operation.
      */
-    public fun executeBatch(expressions: List<SqlExpression>): IntArray {
-        val (sql, _) = formatExpression(expressions[0])
-
-        if (logger.isDebugEnabled()) {
-            logger.debug("SQL: $sql")
-        }
-
-        useConnection { conn ->
-            conn.prepareStatement(sql).use { statement ->
-                for (expr in expressions) {
-                    val (subSql, args) = formatExpression(expr)
-
-                    if (subSql != sql) {
-                        throw IllegalArgumentException(
-                            "Every item in a batch operation must generate the same SQL: \n\n$subSql"
-                        )
-                    }
-                    if (logger.isDebugEnabled()) {
-                        logger.debug("Parameters: " + args.map { "${it.value}(${it.sqlType.typeName})" })
-                    }
-
-                    statement.setArguments(args)
-                    statement.addBatch()
-                }
-
-                val effects = statement.executeBatch()
-
-                if (logger.isDebugEnabled()) {
-                    logger.debug("Effects: ${effects?.contentToString()}")
-                }
-
-                return effects
-            }
-        }
-    }
+    public fun executeBatch(expressions: List<SqlExpression>): IntArray
 
     /**
      * Companion object provides functions to connect to databases.
@@ -536,7 +308,7 @@ public class Database(
             generateSqlInUpperCase: Boolean? = null,
             connector: () -> Connection
         ): Database {
-            return Database(
+            return DatabaseImpl(
                 transactionManager = JdbcTransactionManager(connector),
                 dialect = dialect,
                 logger = logger,
@@ -562,7 +334,7 @@ public class Database(
             alwaysQuoteIdentifiers: Boolean = false,
             generateSqlInUpperCase: Boolean? = null
         ): Database {
-            return Database(
+            return DatabaseImpl(
                 transactionManager = JdbcTransactionManager { dataSource.connection },
                 dialect = dialect,
                 logger = logger,
@@ -598,7 +370,7 @@ public class Database(
                 Class.forName(driver)
             }
 
-            return Database(
+            return DatabaseImpl(
                 transactionManager = JdbcTransactionManager { DriverManager.getConnection(url, user, password) },
                 dialect = dialect,
                 logger = logger,
@@ -632,7 +404,7 @@ public class Database(
             generateSqlInUpperCase: Boolean? = null
         ): Database {
             val translator = SQLErrorCodeSQLExceptionTranslator(dataSource)
-            return Database(
+            return DatabaseImpl(
                 transactionManager = SpringManagedTransactionManager(dataSource),
                 dialect = dialect,
                 logger = logger,
@@ -640,6 +412,83 @@ public class Database(
                 alwaysQuoteIdentifiers = alwaysQuoteIdentifiers,
                 generateSqlInUpperCase = generateSqlInUpperCase
             )
+        }
+    }
+}
+
+/**
+ * Obtain a connection and invoke the callback function with it.
+ *
+ * If the current thread has opened a transaction, then this transaction's connection will be used.
+ * Otherwise, Ktorm will pass a new-created connection to the function and auto close it after it's
+ * not useful anymore.
+ *
+ * @param func the executed callback function.
+ * @return the result of the callback function.
+ */
+@OptIn(ExperimentalContracts::class)
+public inline fun <T> Database.useConnection(func: (Connection) -> T): T {
+    contract {
+        callsInPlace(func, EXACTLY_ONCE)
+    }
+    try {
+        val transaction = transactionManager.currentTransaction
+        val connection = transaction?.connection ?: transactionManager.newConnection()
+
+        try {
+            return func(connection)
+        } finally {
+            if (transaction == null) connection.close()
+        }
+    } catch (e: SQLException) {
+        throw exceptionTranslator?.invoke(e) ?: e
+    }
+}
+
+
+/**
+ * Execute the specific callback function in a transaction and returns its result if the execution succeeds,
+ * otherwise, if the execution fails, the transaction will be rollback.
+ *
+ * Note:
+ *
+ * - Any exceptions thrown in the callback function can trigger a rollback.
+ * - This function is reentrant, so it can be called nested. However, the inner calls don’t open new transactions
+ * but share the same ones with outers.
+ * - Since version 3.3.0, the default isolation has changed to null (stands for the default isolation level of the
+ * underlying datastore), not [TransactionIsolation.REPEATABLE_READ] anymore.
+ *
+ * @param isolation transaction isolation, null for the default isolation level of the underlying datastore.
+ * @param func the executed callback function.
+ * @return the result of the callback function.
+ */
+@OptIn(ExperimentalContracts::class)
+public inline fun <T> Database.useTransaction(isolation: TransactionIsolation? = null, func: (Transaction) -> T): T {
+    contract {
+        callsInPlace(func, InvocationKind.EXACTLY_ONCE)
+    }
+
+    val current = transactionManager.currentTransaction
+    val isOuter = current == null
+    val transaction = current ?: transactionManager.newTransaction(isolation)
+    var throwable: Throwable? = null
+
+    try {
+        return func(transaction)
+    } catch (e: SQLException) {
+        throwable = exceptionTranslator?.invoke(e) ?: e
+        throw throwable
+    } catch (e: Throwable) {
+        throwable = e
+        throw throwable
+    } finally {
+        if (isOuter) {
+            @Suppress("ConvertTryFinallyToUseCall")
+            try {
+                if (throwable == null) transaction.commit() else transaction.rollback()
+            } finally {
+                transaction.close()
+            }
         }
     }
 }
