+++ ktorm-core/src/test/kotlin/org/ktorm/BaseTest.kt
@@ -4,6 +4,7 @@ import org.junit.After
 import org.junit.Before
 import org.ktorm.database.Database
 import org.ktorm.database.use
+import org.ktorm.database.useConnection
 import org.ktorm.entity.Entity
 import org.ktorm.entity.sequenceOf
 import org.ktorm.logging.ConsoleLogger
