+++ ktorm-support-sqlite/src/test/kotlin/org/ktorm/support/sqlite/SQLiteTest.kt
@@ -4,6 +4,7 @@ import org.junit.Test
 import org.ktorm.BaseTest
 import org.ktorm.database.Database
 import org.ktorm.database.use
+import org.ktorm.database.useConnection
 import org.ktorm.dsl.*
 import org.ktorm.entity.count
 import org.ktorm.entity.find
